Crypto/lab3/requirements.txt
argparse
crypto==1.4.1
Pillow==10.0.1
